// shelter-frontend/src/app/api/model/updateZbiorkaDto.ts
/**
 * Shelter
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateZbiorkaDto {
    id?: number;
    nazwa?: string;
    cel?: string;
    kwotaDocelowa?: number;
    kwotaZebrana?: number;
    zwierzeta?: number[];
}